Move get_logo_url from branding/__init__.py to branding/api.py for consistency.
